refactor: merge CLUB_DASHBOARD_NAV_ITEMS and isClubDashboardNavActive into Navbar
